Transfert all logic from AppController & AppService to ReportController & ReportService
