+++ src/app.controller.ts
@@ -1,72 +1,4 @@
-import {
-  Body,
-  Controller,
-  Delete,
-  Get,
-  HttpCode,
-  Param,
-  ParseUUIDPipe,
-  ParseEnumPipe,
-  Post,
-  Put,
-} from '@nestjs/common';
-import { ReportType } from './data';
-import { AppService } from './app.service';
-import {
-  CreateReportDto,
-  ReportResponseDto,
-  UpdateReportDto,
-} from './dto/report.dto';
+import { Controller } from '@nestjs/common';
 
-@Controller('report/:type')
-export class AppController {
-  constructor(private readonly appService: AppService) {}
-
-  @Get()
-  getAlReports(
-    @Param('type', new ParseEnumPipe(ReportType)) type: string,
-  ): ReportResponseDto[] {
-    const reportType =
-      type === 'income' ? ReportType.INCOME : ReportType.EXPENSE;
-    return this.appService.getAllReports(reportType);
-  }
-
-  @Get(':id')
-  getIncomeReportById(
-    @Param('type', new ParseEnumPipe(ReportType)) type: string,
-    @Param('id', ParseUUIDPipe) id: string,
-  ): ReportResponseDto {
-    const reportType =
-      type === 'income' ? ReportType.INCOME : ReportType.EXPENSE;
-
-    return this.appService.getReportById(reportType, id);
-  }
-
-  @Post()
-  createtReport(
-    @Body() { amount, source }: CreateReportDto,
-    @Param('type', new ParseEnumPipe(ReportType)) type: string,
-  ): ReportResponseDto {
-    const reportType =
-      type === 'income' ? ReportType.INCOME : ReportType.EXPENSE;
-    return this.appService.createReport({ amount, source }, reportType);
-  }
-
-  @Put(':id')
-  updateReport(
-    @Param('type', new ParseEnumPipe(ReportType)) type: string,
-    @Param('id', ParseUUIDPipe) id: string,
-    @Body() body: UpdateReportDto,
-  ): ReportResponseDto {
-    const reportType =
-      type === 'income' ? ReportType.INCOME : ReportType.EXPENSE;
-
-    return this.appService.updateReport(reportType, id, body);
-  }
-
-  @HttpCode(204)
-  @Delete(':id')
-  deleteReport(@Param('id', ParseUUIDPipe) id: string) {
-    return this.appService.deleteReport(id);
-  }
-}
+@Controller('')
+export class AppController {}
