+++ src/report/report.controller.ts
@@ -1,4 +1,72 @@
-import { Controller } from '@nestjs/common';
+import {
+  Body,
+  Controller,
+  Delete,
+  Get,
+  HttpCode,
+  Param,
+  ParseUUIDPipe,
+  ParseEnumPipe,
+  Post,
+  Put,
+} from '@nestjs/common';
+import { ReportType } from 'src/data';
+import { ReportService } from './report.service';
+import {
+  CreateReportDto,
+  ReportResponseDto,
+  UpdateReportDto,
+} from 'src/dto/report.dto';
 
-@Controller('report')
-export class ReportController {}
+@Controller('report/:type')
+export class ReportController {
+  constructor(private readonly reportService: ReportService) {}
+
+  @Get()
+  getAlReports(
+    @Param('type', new ParseEnumPipe(ReportType)) type: string,
+  ): ReportResponseDto[] {
+    const reportType =
+      type === 'income' ? ReportType.INCOME : ReportType.EXPENSE;
+    return this.reportService.getAllReports(reportType);
+  }
+
+  @Get(':id')
+  getIncomeReportById(
+    @Param('type', new ParseEnumPipe(ReportType)) type: string,
+    @Param('id', ParseUUIDPipe) id: string,
+  ): ReportResponseDto {
+    const reportType =
+      type === 'income' ? ReportType.INCOME : ReportType.EXPENSE;
+
+    return this.reportService.getReportById(reportType, id);
+  }
+
+  @Post()
+  createtReport(
+    @Body() { amount, source }: CreateReportDto,
+    @Param('type', new ParseEnumPipe(ReportType)) type: string,
+  ): ReportResponseDto {
+    const reportType =
+      type === 'income' ? ReportType.INCOME : ReportType.EXPENSE;
+    return this.reportService.createReport({ amount, source }, reportType);
+  }
+
+  @Put(':id')
+  updateReport(
+    @Param('type', new ParseEnumPipe(ReportType)) type: string,
+    @Param('id', ParseUUIDPipe) id: string,
+    @Body() body: UpdateReportDto,
+  ): ReportResponseDto {
+    const reportType =
+      type === 'income' ? ReportType.INCOME : ReportType.EXPENSE;
+
+    return this.reportService.updateReport(reportType, id, body);
+  }
+
+  @HttpCode(204)
+  @Delete(':id')
+  deleteReport(@Param('id', ParseUUIDPipe) id: string) {
+    return this.reportService.deleteReport(id);
+  }
+}
