+++ src/report/report.module.ts
@@ -4,6 +4,6 @@ import { ReportService } from './report.service';
 
 @Module({
   controllers: [ReportController],
-  providers: [ReportService]
+  providers: [ReportService],
 })
 export class ReportModule {}
